not compile, research
The interface and capabilities for terminal text based interfaces is
limited compared to the multiple proportional font capabilities of a gui
application. Providing a design that encompases terminal for easy
writing such as what is commonly used for utility and command line is
useful. The design of ncurses lends itself well to the discovery of how
text interfaces can be built. With the proper architecture
interoperability between multiple systems and domains is possible.

ncurses
ansi-term
qt5
mozilla
blink/embedded chrome
pango/cairo

One form, runs on each of these.